remove the duplicated file of the project and app fix bearer token problem
