+++ requirements.txt
@@ -1,5 +1,7 @@
 django~=4.1.5
-
+django-seed
 djangorestframework~=3.14.0
-xhtml2pdf~=0.2.9
+xhtml2pdf
 views~=0.3
+psycopg2==2.9.5
+djangorestframework-simplejwt==5.2.2
